+++ PitchForkBanking/src/main/java/com/onlinebanking/services/TransactionService.java
@@ -8,7 +8,7 @@ import com.onlinebanking.models.Transaction;
 
 public interface TransactionService {
 
-	public void createTransaction(String fromAccount, String toAccount, int amount, TransactionType type);
+	public void createTransaction(String fromAccount, String toAccount, double amount, TransactionType type);
 	public void addRequest(Requests request);
 	public List<Requests> getAllRequestsFromUser(String userId);
 	public List<Transaction> getListOfApprovedTransactionRequests(String userId);
