+++ PitchForkBanking/src/main/java/com/onlinebanking/services/TransactionServiceImpl.java
@@ -60,7 +60,7 @@ public class TransactionServiceImpl implements TransactionService {
 	@Override
 	@Transactional
 	public void createTransaction(String fromAccount, String toAccount,
-			int amount, TransactionType type) {
+			double amount, TransactionType type) {
 		Transaction t = new Transaction();
 		Account toAcc = this.accountHome.findById(Integer.parseInt(toAccount));
 		Account fromAcc = this.accountHome.findById(Integer
